Nya mappar views och css
CSS för varje view,
korrigering i App.css och App.jsx pga nya sökvägar
