+++ airbean-gruppovning/src/App.jsx
@@ -2,13 +2,13 @@ import './App.css'
 
 import { RouterProvider, createBrowserRouter, Router } from 'react-router-dom'
 
-import Landing from './views/Landing'
-import Nav from './views/Nav'
-import Menu from './views/Menu'
-import About from './views/About'
-import Cart from './views/Cart'
-import Status from './views/Status'
-import Error from './views/Error'
+import Landing from './views/Landing/Landing'
+import Nav from './views/Nav/Nav'
+import Menu from './views/Menu/Menu'
+import About from './views/About/About'
+import Cart from './views/Cart/Cart'
+import Status from './views/Status/Status'
+import Error from './views/Error/Error'
 
 function App() {
