+++ airbean-gruppovning/src/views/Error/Error.jsx
@@ -1,3 +1,4 @@
+import './Error.css'
 import React from 'react'
 
 function Error() {
