+++ airbean-gruppovning/src/views/Landing/Landing.css
@@ -1,0 +1,20 @@
+
+/* LANDING LANDING LANDING */
+
+.landing-article {
+    width: 375px;
+    height: 667px;
+    display: flex;
+    justify-content: space-between;
+    background-color: #38846D;
+    align-items: center;
+  }
+  
+  .landing-logo {
+    width: 64px;
+    height: 64px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50%;
+  }
