+++ airbean-gruppovning/src/views/Nav/Nav.css
@@ -1,0 +1,59 @@
+/* NAV NAV NAV*/
+
+.nav-article {
+    width: 375px;
+    height: 667px;
+    background-color: #2F2926;
+  }
+  
+  .nav-header {
+    display: flex;
+    justify-content: flex-start;
+    align-items: flex-start;
+  }
+  
+  .nav-section {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 500px;
+    text-align: center;
+  }
+  
+  .roundbox {
+    background-color: lightgray ;
+    margin: 1rem;
+    border-radius: 50%;
+    width: 48px;
+    height: 48px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .cross {
+    max-width: 20px ;
+    max-height: 20px;
+  }
+  
+  ul {
+    list-style-type: none;
+    font-family: 'PT Serif', serif;
+    font-size: 32px;
+    font-weight: 700;
+    line-height: 38px;
+  }
+  
+  li {
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+    color: #fdfdfd;
+  }
+  
+  hr {
+    width: 48px;
+    text-align: center;
+    margin-left: 4rem;
+    color: rgb(59, 59, 59);
+  }
