+++ airbean-gruppovning/src/views/Nav/Nav.jsx
@@ -1,3 +1,4 @@
+import './Nav.css'
 import React from 'react'
 
 function Nav() {
